+++ clients/vscode-hlasmplugin/CHANGELOG.md
@@ -10,6 +10,7 @@
 - Origin of macro statements is tracked incorrectly
 - The language server crashes due to incorrect handling of dependency cycles
 - The language server crashes while processing data constant with cyclic dependency
+- Infinite loop when an invalid EQU operand is found in the lookahead mode
 
 ## [1.15.1](https://github.com/eclipse-che4z/che-che4z-lsp-for-hlasm/compare/1.15.0...1.15.1) (2024-11-22)